refactor: centralize and streamline type definitions for Booking and Room models, removing redundant interfaces and improving code organization
